# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API сервиса автоматического управления ценами на маркетплейсах Wildberries и Ozon.\n\n## Аутентификация\nПосле успешного входа (POST /api/auth/login) сервер устанавливает HttpOnly-cookie `rx_session`.\nБраузер отправляет его автоматически; при работе через curl/Postman передавайте `-b \"rx_session=\u003ctoken\u003e\"`.\n\n## CSRF-защита\nВсе мутирующие эндпоинты (POST/PATCH/DELETE), кроме `/api/auth/register`, `/api/auth/login`\nи `/api/auth/verification/resend`, проверяют заголовок `Origin`.\nОн должен совпадать с одним из разрешённых источников (настраивается через `ALLOWED_ORIGINS`).\nSwagger UI выставляет `Origin` автоматически.",
        "title": "RepricerX API",
        "contact": {
            "name": "Akim Zuev",
            "email": "akim.zuev.86@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Проверяет email и пароль. При успехе создаёт сессию и устанавливает HttpOnly-cookie `rx_session`.\nАккаунт должен быть активирован (email подтверждён через письмо).\nПосле 5 неудачных попыток аккаунт блокируется на 15 минут.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход",
                "parameters": [
                    {
                        "description": "Учётные данные",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход, cookie rx_session выставлен",
                        "schema": {
                            "$ref": "#/definitions/transport.loginResponse"
                        }
                    },
                    "401": {
                        "description": "Неверные учётные данные или аккаунт не активирован (код: invalid_credentials)",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Аккаунт заблокирован (код: user_blocked)",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Завершает текущую сессию, инвалидирует cookie `rx_session`.\nТребует валидной сессии и совпадения заголовка Origin (CSRF-защита).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Выход",
                "responses": {
                    "204": {
                        "description": "Сессия завершена"
                    },
                    "401": {
                        "description": "Сессия не найдена или истекла",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "403": {
                        "description": "CSRF — Origin не совпадает с разрешёнными",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Возвращает данные аутентифицированного пользователя: id, email, displayName, статус, дату регистрации.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Профиль текущего пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.meResponse"
                        }
                    },
                    "401": {
                        "description": "Сессия отсутствует или истекла",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Изменяет отображаемое имя (displayName) текущего пользователя.\nТребует валидной сессии и совпадения Origin (CSRF-защита).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновить профиль",
                "parameters": [
                    {
                        "description": "Новое отображаемое имя",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.updateMeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновлённые данные пользователя",
                        "schema": {
                            "$ref": "#/definitions/transport.loginResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Не аутентифицирован",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "403": {
                        "description": "CSRF — Origin не совпадает",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "description": "Создаёт нового пользователя. На указанный email отправляется письмо для подтверждения.\nВойти можно только после перехода по ссылке из письма (активация аккаунта).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь создан, письмо отправлено",
                        "schema": {
                            "$ref": "#/definitions/transport.registerResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса, некорректный email или слабый пароль (код: bad_request / invalid_email / weak_password)",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "409": {
                        "description": "Email уже зарегистрирован (код: email_taken)",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/verification/resend": {
            "post": {
                "description": "Отправляет новое письмо с токеном верификации.\nВсегда возвращает 202 — сервер не раскрывает, существует ли аккаунт с таким email (защита от перебора).",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Повторная отправка письма",
                "parameters": [
                    {
                        "description": "Email пользователя",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.resendRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Письмо отправлено (или будет отправлено, если email зарегистрирован)"
                    }
                }
            }
        },
        "/api/auth/verify": {
            "get": {
                "description": "Принимает одноразовый токен из письма, активирует аккаунт и перенаправляет на фронтенд.\nПри успехе — редирект на `/login?verified=1`, при ошибке или истёкшем токене — на `/login?verified=0`.",
                "tags": [
                    "auth"
                ],
                "summary": "Подтверждение email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Одноразовый токен из письма (plaintext)",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Редирект на фронтенд с verified=0 (токен не найден, уже использован или истёк)"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "transport.errorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "invalid_credentials"
                },
                "message": {
                    "type": "string",
                    "example": "Неверный email или пароль"
                }
            }
        },
        "transport.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/transport.errorDetail"
                }
            }
        },
        "transport.loginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "MyP@ssword123"
                }
            }
        },
        "transport.loginResponse": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string",
                    "example": "Иван Петров"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "transport.meResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string",
                    "example": "Иван Петров"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "transport.registerRequest": {
            "type": "object",
            "required": [
                "displayName",
                "email",
                "password"
            ],
            "properties": {
                "displayName": {
                    "type": "string",
                    "example": "Иван Петров"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "MyP@ssword123"
                }
            }
        },
        "transport.registerResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
        "transport.resendRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
...
        "transport.updateMeRequest": {
            "type": "object",
            "required": [
                "displayName"
            ],
            "properties": {
                "displayName": {
                    "type": "string",
                    "example": "Новое Имя"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "description": "Сессионный cookie, выставляемый при логине (POST /api/auth/login).",
            "type": "apiKey",
            "name": "rx_session",
            "in": "cookie"
        }
    }
}